refactor tables
